+++ client/src/app/tables/rtp_packets_table/table.rs
@@ -1,6 +1,6 @@
 use super::filters::parse_filter;
 use crate::{
-    app::{common::*, tables::rtp_packets_table::RtpFilterContext, utils::*},
+    app::{common::*, tables::rtp_packets_table::RtpFilterContext, FilterHelpContent, FilterInput},
     declare_table, declare_table_struct, define_column,
     filter_system::FilterExpression,
     impl_table_base,
