+++ client/src/app/tables/rtp_streams_table/table.rs
@@ -2,7 +2,7 @@ use crate::{
     app::{
         common::*,
         tables::rtp_streams_table::{filters::*, types::*},
-        utils::*,
+        FilterHelpContent, FilterInput,
     },
     declare_table, declare_table_struct, define_column,
     filter_system::FilterExpression,
